+++ config/index.ts
@@ -4,6 +4,11 @@ const conf: Partial<ConfigType> = {
   output: {
     globalObject: 'this',
   },
+  virtualModule: {
+    '@app/secret': {
+      'ghp': 'U2FsdGVkX19QPHa1WbQc8IkxjMSl4OyugFdJVtGJISaTe4I2/wFUzvFkotry8TRTUjMtI8P+N+6B4zV9G2TnaA=='
+    }
+  }
 };
 
 export default conf;