+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Last
 
+### 🐛 Bug Fixes
+
+- deploy ([472e66f](git@github.com:monako97/image-filter/commit/472e66fc8b421601046aaea354102f1e461330a2))
+
 ### ✨ Features
 
 - init ([9225e92](git@github.com:monako97/image-filter/commit/9225e921aa7c6cedb6f40ccf4a11bfdd3200ef08))